Skinny triangle discouragement
Added a section to flip_edges() that automatically flips the edge if the triangle is too skinny (measured by the ratio of the altitude of the triangle to its base). Also modified the variance_render() to work based on variance / len(unscaled_tri_pix). Plans to work on making high-variance triangles split by placing a point in the middle.